fix: extract waitForUrl to own module so tests don't pull in @git-glimpse/core
The test was importing from index.ts which transitively imports
@git-glimpse/core, causing Vite resolution failures in CI. Moving
waitForUrl to a standalone module breaks that dependency chain.
