+++ packages/action/src/wait-for-url.ts
@@ -1,0 +1,13 @@
+export async function waitForUrl(url: string, timeout: number): Promise<void> {
+  const deadline = Date.now() + timeout;
+  while (Date.now() < deadline) {
+    try {
+      const res = await fetch(url);
+      if (res.ok) return;
+    } catch {
+      // not ready yet
+    }
+    await new Promise((r) => setTimeout(r, 1000));
+  }
+  throw new Error(`App did not become ready at ${url} within ${timeout / 1000}s`);
+}
